# docs/family-kanban/openapi.yaml
openapi: 3.0.3
info:
  title: Family Scoreboard + Family Kanban — Authorization Skeleton
  version: "0.3.0-module-split"
  description: >-
    這不是完整的 API 規格，只涵蓋授權相關的端點骨架，涵蓋兩個獨立模組：
    /api/family-scoreboard（見 backend/MidoLearning.Api/Endpoints/FamilyScoreboardEndpoints.cs）
    與 /api/family-kanban（見 backend/MidoLearning.Api/Endpoints/FamilyKanbanEndpoints.cs）。
    兩個模組彼此不直接依賴，只共用家庭歸屬授權基礎（IFamilyAccessService +
    FamilyAccessEndpointFilter，見 AGENTS.md 第 3、9 節）。
    目的是讓後續開發者/AI agent 一眼看出「這個端點屬於哪個模組、需要什麼身份、
    通過什麼 gate、失敗會回什麼」。完整欄位規格請直接看對應的 DTO
    （Models/FamilyScoreboard/Dtos.cs、Models/FamilyKanban/Dtos.cs）。
    對應的文字說明見同目錄 AGENTS.md。
servers:
  - url: https://mido-learning-api-24mwb46hra-de.a.run.app
    description: Production
  - url: http://localhost:5199
    description: Local development
security: []
components:
  securitySchemes:
    FirebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Firebase ID Token。FirebaseAuthHandler 驗證後，uid 對應到
        ClaimTypes.NameIdentifier。用於 FamilyAdmin / AuthenticatedOnly /
        SuperAdminOnly policy。
    PlayerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        由 POST /player-login 換發的短期 JWT（type=player claim），
        內含 familyId / playerId / playerName claims。用於 PlayerOnly policy。
    DevApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        僅限 Development 環境的測試後門（見 ApiKey:TestKey 設定）。
        Phase 0 起，非 Development 環境一律停用，即使設定檔誤帶也不生效。
  responses:
    Unauthorized:
      description: 未登入，或 Token 無效/過期。
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: >-
        已登入，但不符合授權條件。可能是：
        (a) FamilyAdmin policy 擋下 player role，或
        (b) CanAccessFamilyAsync 判定 uid 不是該 familyId 的 primary admin / co-admin。
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
...
    PrivateDocDto:
      type: object
      description: >-
        per-user 私密文件（例：保羅寫給配偶的使用說明書）。GET 端點只回傳
        visibleToEmail 與呼叫者 email 相符（大小寫不敏感）的文件，見 AGENTS.md 第 8 節。
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        visibleToEmail:
          type: string
          description: 只有這個 email 的登入者查詢時看得到這份文件。
        createdBy: { type: string, description: 建立者的 uid }
        createdAt: { type: string, format: date-time }
    CreatePrivateDocRequest:
      type: object
      required: [title, content, visibleToEmail]
      properties:
        title: { type: string }
        content: { type: string }
        visibleToEmail: { type: string }
paths:
...
  /api/family-kanban/{familyId}/private-docs:
    post:
      summary: 建立 per-user 私密文件（Parent tier）
      description: >-
        需登入且非 player role（FamilyAdmin policy）+ CanAccessFamilyAsync 通過。
        任何家庭 admin/co-admin 都能建立，visibleToEmail 可指定給家庭內任何人（含自己）。
        見 AGENTS.md 第 8、9 節。
      security:
        - FirebaseIdToken: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrivateDocRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateDocDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      summary: 查詢「我看得到」的私密文件（Parent tier，per-doc email 過濾）
      description: >-
        通過 CanAccessFamilyAsync 只是第一關；回應只包含 visibleToEmail 與呼叫者
        ClaimTypes.Email 相符（大小寫不敏感）的文件。**即使是這個家庭的 primary
        admin，查不到寫給別人的文件也是回 200 + 空陣列，不是 403**——見 AGENTS.md 8.3 節。
      security:
        - FirebaseIdToken: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK（可能是空陣列 —— 代表這個家庭沒有寫給「我」看的文件）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivateDocDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/family-kanban/{familyId}/private-docs/{docId}:
    delete:
      summary: 刪除私密文件（Parent tier）
      description: >-
        任何家庭 admin/co-admin 都能刪除任何一份該家庭的私密文件（沒有 per-doc
        email 限制，見 AGENTS.md 8.5 節已知限制）。
      security:
        - FirebaseIdToken: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema: { type: string }
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ══════════════════════════════════════════════════════════════════════════
  # family-scoreboard 模組（原本的家庭計分板，前綴 /api/family-scoreboard）
  # ══════════════════════════════════════════════════════════════════════════

  # ── PlayerOnly（playerGroup）── 路由 familyId 必須與 JWT familyId claim 相符 ─
...